Add more generic code from OpenOrchestra repo
